+++ CHANGELOG.md
@@ -17,8 +17,9 @@ Possible sections are:
 ### Fixed
 
 - update precomputed average edge lengths (they were underestimated).
-- HUGE improvements, both in  term of speed and memory usage, for
+- HUGE improvements, both in term of speed and memory usage, for
   `Polygon::to_cells`.
+- HUGE improvements, in term of memory usage, for `to_geom` on cells.
 
 ## [0.3.4] - 2023-07-28